refactor(api): Extract health_types.rs (HealthResponse, ComponentHealth)
- Create health_types.rs with 2 DTOs for health checks
- Add 4 unit tests for response serialization
- health.rs: 179 -> 83 lines (-54% reduction)
- Total API lib tests: 339
