+++ edgequake/crates/edgequake-api/src/handlers/mod.rs
@@ -15,6 +15,7 @@ pub mod entities_types;
 pub mod graph;
 pub mod graph_types;
 pub mod health;
+pub mod health_types;
 pub mod lineage;
 pub mod lineage_types;
 pub mod metrics;
@@ -48,6 +49,7 @@ pub use entities_types::*;
 pub use graph::*;
 pub use graph_types::*;
 pub use health::*;
+pub use health_types::*;
 pub use lineage::*;
 pub use lineage_types::*;
 pub use metrics::*;
